Refactored Sliders out of menu.
