+++ src/main/java/roborally/gameview/menu/Menu.java
@@ -12,9 +12,7 @@ import com.badlogic.gdx.scenes.scene2d.Actor;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.Stage;
-import com.badlogic.gdx.scenes.scene2d.ui.Image;
-import com.badlogic.gdx.scenes.scene2d.ui.Label;
-import com.badlogic.gdx.scenes.scene2d.ui.Slider;
+import com.badlogic.gdx.scenes.scene2d.ui.*;
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener;
 import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
 import roborally.game.IGame;
@@ -28,21 +26,6 @@ import java.util.HashMap;
 
 // Menu under construction!
 public class Menu {
-    public static float dt = 0.00000000000001f;
-    private ArrayList<Slider> sliders;
-    private Slider.SliderStyle sliderStyle1;
-    private float[] decoIterate;
-    private int[] decoScale;
-    private float[] decoWidth;
-    private float superit;
-    private float superit2;
-    private boolean flip1;
-    private boolean flip2;
-    private boolean decoH1;
-    private boolean decoH2;
-    private boolean resume;
-    private boolean flip3;
-    private boolean decoH3;
     private boolean changeMap;
     private Music Song;
     private int mapId;
@@ -60,22 +43,19 @@ public class Menu {
     private Image mapButton;
     private Image continueButton;
     private Events events;
+    private boolean resume;
     private int gSpeed;
     private int lSpeed;
     private boolean Continue;
+    private Sliders sliders;
 
     public Menu(Stage stage, Events events) {
         labelStyle = new Label.LabelStyle();
         labelStyle.font = new BitmapFont();
-        this.decoIterate = new float[3];
-        this.decoScale = new int[3];
-        this.decoWidth = new float[3];
         this.resume = false;
         this.changeMap = false;
         this.mapId = 1;
-        this.superit = 0;
         this.events = events;
-        this.sliders = new ArrayList<>();
         imageLists = new HashMap<>();
         startGame = 0;
         imageLists.put("menus", new ArrayList<>());
@@ -85,59 +65,23 @@ public class Menu {
         mapButton = new Image(AssetManagerUtil.getMenu().getMapButton());
         mapButton.setPosition(680, 480);
         changeMapMenu = false;
-        this.decoScale[0] = 2;
-        this.decoScale[1] = 2;
-        this.decoScale[2] = 2;
         addEvenMoreListeners();
         makeLabels();
         makeOptionListeners();
         addMaps();
         addMoreListeners();
-        stage.addActor(continueButton);
         stage.addActor(gameSpeed);
         stage.addActor(laserSpeed);
         stage.addActor(volumeSlider);
         stage.addActor(playSong);
         makeClickableButtons(stage);
+        this.sliders = new Sliders();
+        sliders.abc();
+
     }
 
     private void addEvenMoreListeners() {
-        Pixmap colur = new Pixmap(100, 10, Pixmap.Format.RGBA8888);
-        Pixmap colur2 = new Pixmap(10, 10, Pixmap.Format.RGBA8888);
-        Pixmap colur3 = new Pixmap(2, 2, Pixmap.Format.RGB888);
-        Pixmap colur4 = new Pixmap(10, 1, Pixmap.Format.RGBA8888);
-        colur3.setColor(Color.CHARTREUSE);
-        colur3.fill();
-        colur4.setColor(Color.ORANGE.set(0, 0, 0, 0));
-        colur4.fill();
-        Slider.SliderStyle sliderStyle = new Slider.SliderStyle();
-        sliderStyle1 = new Slider.SliderStyle();
-        colur.setColor(Color.BLACK);
-        colur.fill();
-        colur2.setColor(Color.WHITE);
-        colur2.fill();
-        sliderStyle.background = new Image(new Texture(colur)).getDrawable();
-        sliderStyle.knob = new Image(new Texture(colur2)).getDrawable();
-        sliderStyle1.background = new Image(new Texture(colur4)).getDrawable();
-        sliderStyle1.knob = new Image(new Texture(colur3)).getDrawable();
-        Slider decor1 = new Slider(0, 10, 0.01f, false, sliderStyle1);
-        decor1.setSize(170, 0.2f);
-        decor1.setPosition(420, 350);
-        sliders.add(decor1);
-        decoWidth[0] = 140;
-        Image birdImage = new Image(new Texture("assets/icons/Icon.png"));
-        birdImage.setSize(4, 4);
-        sliderStyle1.knob = birdImage.getDrawable();
-        Slider decor2 = new Slider(0, 10, 0.01f, false, sliderStyle1);
-        decor2.setSize(225, 0.2f);
-        decor2.setPosition(405, 485);
-        sliders.add(decor2);
-        Slider decor3 = new Slider(0, 10, 0.01f, false, sliderStyle1);
-        decor3.setSize(180, 0.2f);
-        decor3.setPosition(430, 420);
-        sliders.add(decor3);
-        decoWidth[1] = decor2.getWidth();
-        decoWidth[2] = decor3.getWidth();
+
         continueButton = new Image(new Texture(Gdx.files.internal("menu/continue.png")));
         continueButton.setPosition(350, 490);
         continueButton.setWidth(continueButton.getWidth() + 150);
@@ -223,7 +167,7 @@ public class Menu {
             @Override
             public boolean touchDown(InputEvent event, float x, float y, int pointer, int button) {
                 if (mapId == 0)
-                    mapId = 2;
+                    mapId = 1;
                 else
                     mapId -= 1;
                 return true;
@@ -232,7 +176,7 @@ public class Menu {
         right.addListener(new InputListener() {
             @Override
             public boolean touchDown(InputEvent event, float x, float y, int pointer, int button) {
-                if (mapId == 2)
+                if (mapId == 1)
                     mapId = 0;
                 else
                     mapId += 1;
@@ -244,6 +188,11 @@ public class Menu {
             @Override
             public void enter(InputEvent event, float x, float y, int pointer, Actor fromActor) {
                 playSong.setColor(Color.GREEN);
+            }
+
+            @Override
+            public void exit(InputEvent event, float x, float y, int pointer, Actor toActor) {
+                playSong.setColor(Color.WHITE);
             }
 
             @Override
@@ -311,10 +260,6 @@ public class Menu {
         return this.mapId;
     }
 
-    public void setMapId(int mapId) {
-        this.mapId = mapId;
-    }
-
     public Events getEvents() {
         return this.events;
     }
@@ -328,14 +273,9 @@ public class Menu {
             stage.addActor(image);
         stage.addActor(gameSpeed);
         stage.addActor(laserSpeed);
-        stage.addActor(continueButton);
         stage.addActor(volume);
         stage.addActor(volumeSlider);
         stage.addActor(playSong);
-    }
-
-    public void setStartGame() {
-        this.startGame = 0;
     }
 
     public void drawMenu(SpriteBatch batch, Stage stage) {
@@ -348,7 +288,7 @@ public class Menu {
             volumeSlider.draw(batch, 1);
             volume.draw(batch, 1);
             playSong.draw(batch, 1);
-            drawSliders(batch);
+            sliders.drawSliders(batch);
             if (startGame == 1)
                 continueButton.draw(batch, 1);
         } else {
@@ -362,46 +302,8 @@ public class Menu {
         stage.act();
     }
 
-    private void drawSliders(SpriteBatch batch) {
-        for (int i = 0; i < sliders.size(); i++) {
-            sliders.get(i).draw(batch, 1);
-            superit2 -= Gdx.graphics.getDeltaTime() * 10;
-            if (superit < 1) {
-                superit += Gdx.graphics.getDeltaTime() * 10;
-                sliders.get(i).setValue(decoIterate[i] += Gdx.graphics.getDeltaTime() * decoScale[i] * 2);
-            } else if (superit2 < -10) {
-                superit = superit2;
-                superit2 = 0;
-            }
-            if (superit >= 1) {
-                sliders.get(i).setValue(decoIterate[i] += Gdx.graphics.getDeltaTime() * decoScale[i] / 2);
-                superit -= Gdx.graphics.getDeltaTime() * 10;
-            }
-            checkForFlips(i);
-        }
-    }
-
-    private void checkForFlips(int i) {
-        if (decoH1 && i == 0 || decoH2 && i == 1 || decoH3 && i == 2) {
-            if (!flip1 && i == 0 || !flip2 && i == 1 || !flip3 && i == 2) {
-                if (decoIterate[i] < sliders.get(i).getMinValue() + dt) {
-                    flipH(i);
-                }
-            }
-            if (flip1 && decoIterate[i] > sliders.get(i).getMaxValue() - dt && i == 0 || flip2 && decoIterate[i] > sliders.get(i).getMaxValue() - dt && i == 1
-                    || flip3 && decoIterate[i] > sliders.get(i).getMaxValue() - dt && i == 2) {
-                flipH2(i);
-            }
-        } else {
-            if (!flip1 && i == 0 || !flip2 && i == 1 || !flip3 && i == 2) {
-                if (decoIterate[i] > sliders.get(i).getMaxValue() - dt) {
-                    flipV(i);
-                }
-            }
-            if (flip1 && decoIterate[i] < sliders.get(i).getMinValue() + dt && i == 0 || flip2 && decoIterate[i] < sliders.get(i).getMinValue() + dt && i == 1 || flip3 && decoIterate[i] < sliders.get(i).getMinValue() + dt && i == 2) {
-                flipV2(i);
-            }
-        }
+    public void setStartGame() {
+        this.startGame = 0;
     }
 
     private void makeLabels() {
@@ -417,12 +319,12 @@ public class Menu {
         playSong.setPosition(750, 640);
         left = new Label(" < ", labelStyle);
         right = new Label(" > ", labelStyle);
-        left.setSize(120, 120);
-        right.setSize(120, 120);
-        left.scaleBy(2);
-        right.scaleBy(2);
+        left.setFontScale(2);
+        right.setFontScale(2);
         left.setPosition(40, 300);
         right.setPosition(665, 300);
+        left.scaleBy(2);
+        right.scaleBy(2);
     }
 
     private void makeClickableButtons(Stage stage) {
@@ -458,7 +360,6 @@ public class Menu {
                 stage.clear();
                 stage.addActor(gameSpeed);
                 stage.addActor(laserSpeed);
-                stage.addActor(continueButton);
                 stage.addActor(volume);
                 stage.addActor(volumeSlider);
                 stage.addActor(playSong);
@@ -527,86 +428,12 @@ public class Menu {
         ArrayList<Image> maps = new ArrayList<>();
         maps.add(new Image(new Texture(Gdx.files.internal("maps/models/map0.png"))));
         maps.add(new Image(new Texture(Gdx.files.internal("maps/models/map1.png"))));
-        maps.add(new Image(new Texture(Gdx.files.internal("maps/models/map2.png"))));
         for (Image image : maps)
             image.setPosition(75, 115);
         imageLists.put("maps", maps);
     }
 
-    public void flipV(int i) {
-        Slider slider = new Slider(0, 5, 0.1f, true, sliderStyle1);
-        slider.setSize(10, 30);
-        slider.setPosition(sliders.get(i).getX() + sliders.get(i).getWidth() - slider.getWidth(), sliders.get(i).getY() - slider.getHeight());
-        slider.setValue(5 - dt);
-        decoIterate[i] = 5 - dt;
-        decoScale[i] = -decoScale[i];
-        sliders.set(i, slider);
-        if (i == 0)
-            decoH1 = true;
-        if (i == 1)
-            decoH2 = true;
-        if (i == 2)
-            decoH3 = true;
+    public void addContinueButtonToStage(Stage stage) {
+        stage.addActor(continueButton);
     }
-
-    public void flipH(int i) {
-        Slider slider = new Slider(0, 10, 0.01f, false, sliderStyle1);
-        slider.setSize(decoWidth[i], 0.2f);
-        slider.setPosition(sliders.get(i).getX() - slider.getWidth() + sliders.get(i).getWidth(), sliders.get(i).getY());
-        slider.setValue(10 - dt);
-        decoIterate[i] = 10 - dt;
-        sliders.set(i, slider);
-        if (i == 0) {
-            decoH1 = false;
-            flip1 = true;
-        }
-        if (i == 1) {
-            decoH2 = false;
-            flip2 = true;
-        }
-        if (i == 2) {
-            decoH3 = false;
-            flip3 = true;
-        }
-    }
-
-    private void flipV2(int i) {
-        Slider slider = new Slider(0, 5, 0.1f, true, sliderStyle1);
-        slider.setSize(10, 30);
-        slider.setPosition(sliders.get(i).getX(), sliders.get(i).getY());
-        slider.setValue(0 + dt);
-        decoIterate[i] = 0 + dt;
-        decoScale[i] = Math.abs(decoScale[i]);
-        sliders.set(i, slider);
-        if (i == 0)
-            decoH1 = true;
-        if (i == 1)
-            decoH2 = true;
-        if (i == 2)
-            decoH3 = true;
-    }
-
-    private void flipH2(int i) {
-        Slider slider = new Slider(0, 10, 0.01f, false, sliderStyle1);
-        slider.setSize(decoWidth[i], 0.2f);
-        slider.setPosition(sliders.get(i).getX(), sliders.get(i).getY() + sliders.get(i).getHeight());
-        slider.setValue(0 + dt);
-        sliders.set(i, slider);
-        decoIterate[i] = 0 + dt;
-        decoScale[i] = Math.abs(decoScale[i]);
-        if (i == 0) {
-            decoH1 = false;
-            flip1 = false;
-        }
-        if (i == 1) {
-            decoH2 = false;
-            flip2 = false;
-        }
-        if (i == 2) {
-            decoH3 = false;
-            flip3 = false;
-        }
-
-    }
-
 }
